refactor: その他コンポーネントについて、componentsフォルダに格納
