+++ src/features/forecast/components/PortionIndicator.tsx
