+++ src/features/hint/components/AlertItem.tsx
